[minfs] Remove dirty cache conditional.
The dirty cache should always be used on devices so this removes the
build flag and the logic around handling it.

Keeps the dirty cache virtual function on Vnode to avoid changing
behavior. The dirty cache is still disabled for directories and on host
builds.

There should be no behavior change.

Change-Id: I2e2be2df57617690662293826d1bf998aadd1268
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610864
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>