[virtcon] Set TERM=xterm-256color for virtcon
Terminfo is commonly available for this terminal type and it
matches the escape code support in the current implementation.

The Alacritty terminal that uses the same term-model also sets
TERM to xterm-256color.

This improves application behavior. E.g. top utility understands
that output is a terminal and not serial.

Change-Id: Ia58027573c8b3f606a3c6683d90f797d2b0466c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611481
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: David Reveman <reveman@google.com>