[realm_builder] compare resolver decls against pkgfs decls
Some builds automatically insert a use declaration for the DebugData
protocol into component manifests. This is an issue in tests that are
expecting to know the contents of the manifests on disk. Work around
this by having the tests compare the contents of the realm builder
resolver against the manifests as they exist in pkgfs, instead of
comparing then against hardcoded values.

Fixed: 89446
Change-Id: I75ee3ecfe12fb66581b83efba3262eb13db27f02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609862
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Commit-Queue: Derek Gonyeo <dgonyeo@google.com>