[settings] match watch to watch2
As part of the migration process for settings.Input, the clients
currently on Watch2 and SetStates will be migrated to the more
properly named Watch and Set. For this to happen, the underlying
functionality of Watch and Set will need to match the current Watch2 and
SetStates. Watch2 and SetStates will be removed after clients are
migrated.

Bug: 65686
Test: fx test setui_service_tests, fx test setui_client_interface_tests

Change-Id: I8a3f6a3383541d2f73bbf5eee719cbc02246ddbc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/603143
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
API-Review: Kevin Lindkvist <lindkvist@google.com>