[ffx][config] Don't save if config did not change
There is no locking between the ffx daemon and the ffx CLI, and so they
can race each other when saving to the config. To avoid that, this patch
changes ffx config when setting a value to only save the config if we
detect that the value has actually changed. This will reduce the writes
to the config, and lower the chance that a race will occur.

Bug: 81502
Change-Id: I42f416027164a25c79205eecffe299fdd6451dbc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610773
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Andrew Davies <awdavies@google.com>