[overnet] Wake quiche writers that return Err(Done)
When quiche returns Err(Done) from stream_send(), it may not have
actually established the stream. This can mean that when we call
writable() on the connection, we won't ever get that stream ID.

Since we go to sleep when we get Err(Done), and we use writable() to
determine when we wake up again, this means writing stalls forever.

To fix this, we add a new set of wakers that is always woken when the
connection becomes writable without checking if a particular stream is
writable. Err(Done) adds us to that waker set rather than the usual
waker set for writers.

This may fix a few connection stall issues we've seen previously, and is
absolutely necessary for the Quiche 0.10 upgrade.

Bug: 79699
Bug: 87439
Bug: 87372
Change-Id: I27a7389cef81527a3ccec7d28c4acd574105ea1e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611746
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>