[ffx] Add a few behavior definitions for cpu/kvm
This CL introduces Behaviors for Femu. These are configuration snippets
that the emulator engine uses to build the command line. A Behavior is
defined by its data field and the type which implements that Behavior.
For Femu, the data field includes four properties: args, features,
kernel_args, and options. The implementing type will parse these
properties and inject them at the correct point in the emulator command
line, assuming the filter for that type is accepted.

This CL doesn't introduce the implementing types yet, or any of the
infrastructure for utilizing the Behaviors. These come soon. This just
provides a few basic behaviors to build on: one simple Behavior, which
is always accepted and included, and a few related complex Behaviors,
which will do some filtering based on runtime flags to select the
correct one.

Change-Id: I5a6cb8324d4f652b6c6034f0638a43945fe6ab4e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611748
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>