[ui] Move terminal rendering logic to a library
This library will be used to share code between virtcon
and the Terminal app.

This also fixes some cursor rendering problems that were
a result of assuming that the cursor and cell content
would not occupy the same cell.

Includes a new set of a unit tests this rendering code.
This was made possible by making some small adjustments
to carnelian to reduce the complexity of creating a Mold
context for testing.

Constants were also added to carnelian that makes it
easier to determine how to limit the size of the terminal
in order to not exceed the current layer limit.

Bug: 74628
Test: terminal_lib_tests
Change-Id: Ibb2f618a6083d39eb09640d3d8d4d125017aa5ec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610386
Commit-Queue: David Reveman <reveman@google.com>
Reviewed-by: Rob Tsuk <robtsuk@google.com>