[build] Check for Go files missing from `sources`
This would have prevented the problems that I fixed in fxrev.dev/604261
and fxrev.dev/603765 where we had forgotten to hook some Go files into
the build graph and some tests had bitrotted.

We also now enforce that all `sources` are in the root of `source_dir`
rather than in a subdirectory. There's no reason to specify `sources`
from a subdirectory of the BUILD.gn file's directory because
`source_dir` can be set to the subdirectory instead. This removes an
unnecessary degree of freedom from setting up a go_library, and also
makes the gen_library_metadata logic a bit simpler because it can assume
`sources` are all in the root of `source_dir`.

Change-Id: I7de17d0091449a1c2e6f379a15b7d7a3765bbb65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/606127
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>