[memalloc] Introduce Pool::size()
This change introduces an O(1) size accessor for memalloc::Pool,
returning the number of tracked, normalized ranges. This will be useful
for when we get into the business of allocating buffers whose size will
be bounded by or be a function of the size of these ranges (e.g., with
physboot passing coalesced ranges in its hand-off).

Change-Id: I20aad043855d6cda34a16554367fea06b16ae6a9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611723
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>