[rust-3p-crates] Bump rustls to 0.19.1
Updating rustls is a little challenging, since the latest version of
hyper-rustls and tokio-rustls depend on the latest versions of hyper and
tokio, which are quite large.

Fortunately we can reduce our load, because all the rustls dependencies
don't actually depend on any rustls breaking API changes. So this patch
takes rustls 0.19.1, updates it to incorporates the fix for
https://fxrev.dev/479224, and resets the version back to 0.17.0.

We also get to remove base64 0.12.3.

Each individual step has been reviewed in:

* 0.17.0 -> 0.18.1: https://fxrev.dev/577407
* 0.18.1 -> 0.19.0: https://fxrev.dev/577408
* 0.19.0 -> 0.19.1: https://fxrev.dev/577409
* Restore https://fxrev.dev/479224: https://fxrev.dev/574590

Bug: 87798

Change-Id: I7ec3977583b6f73680b7e44944321789d1b6dc66
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610999
Reviewed-by: Kostya Kortchinsky <kostyak@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>