[x86][acpi][ec] Only acquire global lock if it's necessary.
Previously, the EC driver contained an assumption that _all_ requests to
the EC required the global lock. This is only the case if the "_GLK"
method exists on the EC device and it evaluates to 1 (ACPI v6.4, section
12.8 says "An embedded controller will require the inclusion of the GLK
method in its ACPI namespace if potentially contentious accesses to
device resources are performed by non-OS code.").

This CL modifies the EC driver to only acquire the global lock if the
_GLK method exists and evaluates to 1.

Bug: 87281
Test: boot atlas and check EC-related functionality still works.
Change-Id: Ida7311b2c4088d642515235bd46cbb56c3cc2bbb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611882
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>