[ffx] Add rust types for emulation Behaviors
This CL builds on the previous one by adding the container types for
Femu Behaviors. These are the data types that store the contents of the
manifest's Behaviors after deserialization. We also update the test
which validates the manifest to include an example behavior.

Change-Id: I2060e2e07e6f3413d4589ee6dc048aa4abff9546
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611808
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>