[zircon][vm][ppb] Fix a couple allocation failure paths
OnClose() can get called with phys_base_ still kInvalidPhysBase
if allocation fails.  This one is more likely to get hit because
the problematic failure path can be taken due to bulk page alloc
failure.

Similarly, but less likely to be hit, the destructors of
PhysicalPageProvider and PagerProxy can be called with closed_
false in allocation failure paths.  This isn't as likely because
this won't happen due to bulk page alloc failure, only due to
failure to allocate a small object.

BUG: fxbug.dev/89538
Change-Id: I798cd66bf5d01e758a2f07865010aae8902d826f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612489
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>