[fdio] Do not allocate 65k on the stack
In commit 299662b6df7da08a178792d8951d788c23138877 fdio was migrated
from a struct-based FIDL API to a table-based one; one of the
consequences of moving to tables is that FIDL clients must always
allocate the maximum size of a channel payload before reading, which is
pathological when the allocation takes place on the stack, as it does in
fdio's getaddrinfo.

Remove the stack buffer so that allocation happens on the heap instead
of occasionally stack smashing.

Bug: 89478
Change-Id: I71160781b7b6ae8e15e2a9dc14129ca30c8cb1e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612421
Commit-Queue: Tamir Duberstein <tamird@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>