[wlan] Create a components library for drivers
This change creates an initial set of basic primitives usable in wlan
drivers that implement the generic network device interface. These
components initially center around the basic transmissible, the frame,
and its storage and lifetime management.

A frame is a collection of metadata relevant to transmission and
reception of frames. Notably it does not contain, allocate or manage the
lifetime of any memory containing the frame data. Instead it contains a
VMO identifier and a pointer to where that VMO has been mapped into
memory.

In addition to a frame object there is a class that manages storage and
lifetime for frames and a class for passing around a collection of
frames.

Bug: 79622
Test: Included unit tests
Change-Id: I87449acc4eb8c6f44db57925281447711d5fa329
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/561101
Commit-Queue: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: John Sheu <sheu@google.com>