[vmm][virtio-sound] Add a Sequencer type
This CL adds a Sequencer type which can be used to create a sequence
of async tasks. This will be used by a future CL that implements audio
packet forwarding:

1. Packets must be forwarded in the same order they are popped from
   a virtqueue, however
2. Mulitple packets may be in flight at any time

This can be done by spawning an async task per packet, where each task
uses a `sequencer::Ticket` to ensure its packet is forwarded in the
correct order.

Sequencer is implemented as a doubly-linked lists of tokio::sync::Notify
objects. Each ticket references a list node. A task acquires its spot in
the sequence by waiting on a Notify object, after which point the Ticket
is "locked". When the task has forwarded its audio packet, it releases
its Ticket lock, which notifies the next Ticket in the sequence.

Bug: 87645
Tested: added unit tests
Change-Id: Iabe509933432f6b2424fd772bec1ffce6842e4f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612413
Commit-Queue: Tom Bergan <tombergan@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>