[wlan][mlme][wlanif] Scan request fixes for channel and SSID lists
The duplicate ssid field is removed in favor of a single ssids_list
field for scan requests. Also, the channel and SSID lists are no
longer arbitrarily limited in mlme and wlanif.

This change also makes it an error to submit an empty channel list in
a scan request. That behavior is not currently used since SME always
submits a non-empty list in practice, so this does not result in a
behavioral change.

Bug: 88649, 88650
Test:
1. fx test wlan-sme-tests wlan-mlme-tests \
  fuchsia-pkg://fuchsia.com/wlanif_unittest#meta/device_test.cm \
  fuchsia-pkg://fuchsia.com/brcmfmac_sim_test#meta/escan_params_test_component.cm \
  fuchsia-pkg://fuchsia.com/brcmfmac_sim_test#meta/passive_scan_test_component.cm \
  fuchsia-pkg://fuchsia.com/brcmfmac_sim_test#meta/active_scan_test_component.cm
2. fx test //src/connectivity/wlan
3. Manual save and connect to network on Astro succeeds.
4. ACTS smoke tests pass
5. Manual save and connect to network on Atlas succeeds.

Change-Id: I1d6d6307084f8da7bd9fa6727e1767ef056aed43
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605366
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Reviewed-by: Joshua Conner <joshconner@google.com>
Reviewed-by: Rebecca Silberstein <silberst@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>