[fidl] Refactor JSON IR to store method payloads as types
Previously, the JSON IR stores "maybe_request\response_payload"
as a simple string pointing to the underlying struct type of the
payload. As we move to using other layouts (namely unions or
tables) as top-level method request/response payloads, this will
be insufficient for most backends, as they will need to
introspect about the underlying type of the payload, rather than
just assuming it is a Struct populated by StructMembers.

This is the first of several changes refactoring the JSON IR,
with the eventual goal being the removal of the (deprecated, but
still widely used) "maybe_request/response" fields, and the
removal of "maybe_success_variant_struct," which will be
unnecessary in a world where all payload shape information is
stored as types rather than structs.

Test: fidlde regen fidlc
Test: fidldev test all
Change-Id: Ic9a9760e86ba344d4b1e2f8ad40dab417fe73501
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605061
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>