[bt][test-harness] Migrate bt-host-integration-tests to CFv2
In order to do so, create a CFv2-based HostDriverHarness. The CFv2
HostDriverHarness operates on a HostDriverRealm instead of a CoreRealm,
like the rest of the CFv2 bt-test-harnesses, because it requires direct
access to the bt-host driver, whereas the other harnesses operate on
services provided by bt-gap/bt-init. The v2 HostDriverRealm is simply
a driver test realm with a bt-hci-emulator device and a bt-host device.

Aside from simply changing the bt-host-integration-tests to use the v2
HostDriverHarness, one test operated directly on the HCI emulator to
verify the bt-host lifecycle. This test has been converted to use
the new HostDriverRealm to do this testing.

MULTIPLY: bt-host-integration-tests (fuchsia): 100
Bug: 53915, 89132
Test: fx test bt-host-integration-tests
Change-Id: Icf5367fcc904ff302340bfdbcf1925d9d8e168ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610548
Fuchsia-Auto-Submit: Lucas Jenkins <lucasjenkins@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Lucas Jenkins <lucasjenkins@google.com>