[ui][input-pipeline] Handlers should only match non-`handled` events (2/2)
This change modifies the following handlers to only handle input events with a `handled` property of `input_device::Handled:No`:
- gfx_touch_handler
- ime_handler
- media_buttons_handler
- mouse_handler

This set of handlers was chosen because their unit tests share in listening to a request stream to verify input events are received and acted upon. In order to assert that handled events are not received by a request stream, a common testing utility `assert_handler_ignores_input_event_sequence` is created and shared among them.

This is a counterpart to https://fuchsia-review.googlesource.com/c/fuchsia/+/610389

Bug: 73818
Change-Id: Idbd46e91333b060392e1393c6cd65ac1fa5e661b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611781
Commit-Queue: Caroline Liu <carolineliu@google.com>
Reviewed-by: Filip Filmar <fmil@google.com>