[kernel][phys] Consume UART driver information from hand-off
This change hands off the UART driver information from physboot to the
kernel, in the form of uart::all::Driver. While the kernel is not yet
able to actually use the driver as handed off, it is able to make direct
use of the encoded, associated metadata, forgoing the ZBI parsing
otherwise needed to obtain this info.

Bug: 88112,89182
Change-Id: I245b724a33c6fe2911a269496d2c149b3282fe8b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609721
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>