[build][drivers] Enforce drivers info file inclusion
This CL makes it a hard build failure for a public driver (defined
as drivers in //src and //zircon) to not provide an info file.
An info file is automatically created by fx create driver or it can
be manually added.

pshickel@ suggested we document this requirement in
fuchsia_driver_component.gni, so we add comments stating this.

Test: Removing an info file for a driver in //src results in a build assertion reporting "Must include driver info"
Fixed: 81799
Change-Id: I86046a15d53adec39e8c8a99decd8861b00504c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612132
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>