[fidl][cpp] Minimal struct domain object
The intention of this CL is to support using a standalone struct as
requests/responses. Introducing a natural struct allows us to generate
request/response payloads as regular domain objects in a later CL.

The example code was updated to demonstrate some of the niceties in the
new API, such as accessors and designated initialization (a la named
arguments).

fxbug.dev/82189 is a bigger umbrella ticket to fully define the whole
set of natural domain objects.

Test: fx test //src/lib/fidl/cpp

Bug: 60240
Bug: 82189
Change-Id: Ib395cbbb4b2261ff0b8e94755bd68d437ac28664
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612982
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>