[vmm] Avoid time-of-check/time-of-use race in IO APIC
The IO APIC currently takes/releases its internal lock twice when
writing IRQ redirect entries. The first time is in `Write` to determine
which indirect register is selected, and then again in `WriteRegister`
when attempting to write the actual entry.

Unfortunately, the current code has a potential race if `select_`
changes between the two locks:

    switch (select_register) {   // ** original value of `select_` **
      ...
      case kFirstRedirectOffset ... kLastRedirectOffset: {
        ...
        // *** new value of `select_` ***
        uint32_t redirect_offset = select_ - kFirstRedirectOffset;
        RedirectEntry& entry = redirect_[redirect_offset / 2];

In the switch statement, we verify that the original value of `select_`
is in the range `kFirstRedirectOffset` to `kLastRedirectOffset`.
However, we then use the new value of `select_` to index into the array
`redirect_`.

Because we drop the mutex between the first and second checks, the risk
is that the value of `select_` may have changed between the two calls,
meaning we access an out-of-bounds array element.

The `Read` path does the correct action here, performing the second
calculation with the original value of `select_`.

This CL updates the code to just hold onto the mutex for the entire
Read/Write: currently, we only drop it for a few cycles, and the
overhead of doing the drop/acquire will outweigh the benefit of holding
it for a shorter amount of time.

Change-Id: I627d5a39e30e2c9ee42bf0febe102781c6f04d23
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613544
Commit-Queue: David Greenaway <dgreenaway@google.com>
Fuchsia-Auto-Submit: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>