[wlan][mlme][wlanif] Impose scan request channels limit and update docs
Channel numbers in IEEE 802.11-2016 fundamentally limited to a single
octet, so it's unnecessary to allow specifying more than 255 non-zero
channels in a single scan request.

Documentation on the SSID list is updated to reflect the technical
limitation which would make servicing a scan request with a large
number of SSIDs difficult.

Bug: 88648
Test: None. New channel list limit cannot cause otherwise valid scan
requests to now fail.

Change-Id: Ia6bc600fa5faada4aebbc728351eccb73bd10d68