[component_manager] reduce log level in storage admin
The storage admin protocol will issue a warning log if it fails to lock
the resolved state of a component. This situation can be handled
gracefully, and may be occasionally encountered during component
destruction, so the `warn` log level is too high. Reduce the log level
to `debug`, so as not to alarm users who see the log.

Change-Id: I96f1f560fae52b199ad2f0ee59cc290a3af50bd1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613643
Fuchsia-Auto-Submit: Derek Gonyeo <dgonyeo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>