[licenses] Allow unknown json fields in the config file.
I'd like to make it easier for anyone to run fx check-licenses.
Instead of having to provide lengthy paths to the config files on
the command line, I would like it to do the right thing by default.

A strategy I landed on is to create a new field in the config file:
"Include", which is a list of paths to other config files. Every file
in that list will be merged with the original config file, just like they
are when you list them out on the command line.

Explicit implementation details will come in future CLs, but I need to
allow unknown json fields so I can make the change in //vendor/google,
so this whole process is a soft transition.

Change-Id: I3e5b29833c316dc8d979f4dbb39b0770ddc07e76
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613474
Reviewed-by: Sean Cuff <seancuff@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>