[ffx] Add behavior handlers for ffx emu
This CL adds the behavior trait and a few implementations of it to ffx
emu. These are the Rust code that are referred to by the JSON behavior
specs, and implement the processing of those behaviors. The
femu_behavior file includes generic handling of the data properties for
that emulator by providing access to the appropriate fields as methods
on the data type. The simple_behavior file extends that by combining it
and the BehaviorTrait, which provides a no-op version of the
setup/cleanup/filter methods. These are the default behavior handlers
for those engines.

We also introduce a few complex behavior handlers for the KVM behaviors.
These handlers still pull the data from the same fields, so they extend
femu_behavior accordingly. But different filtering code is needed for
each handler, so they reimplement BehaviorTrait accordingly.

All that, plus an enumeration for mapping the Rust handlers to the JSON
strings in the handler fields.

Change-Id: I70a383003339946570fd5d5f0c50b5d290eb8a15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611943
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>