[ffx][test] Verify structured output conforms with schema in tests
It's easy to update the format without updating the schema. This adds
a check against the schema whenever the output format is deserialized
in tests. This technically violates the locality principle outlined by
the testability rubric, since we'll be validating the 'directory' crate
using tests in 'run_test_suite', but it will help us catch in the
future if a test case in 'directory' is missing.

Change-Id: I1321e3e65924797a7e3b3855c1f5b2b53c94c06b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610867
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>