[input] Adds LockState to the KeyEvent API
`LockState` maintains the state of the lock keys, for cross-runner
consistency.

`Modifiers` maintain the actuation state of the keys.

Added the definitions of some common modifiers and lock states
so that we can implement a handler that maintains those.

Shortened the imports to reduce the visual clutter in the code.

The modifiers are not yet forwarded to clients.  Hooking them up is
relegated to a future commit, since this one is already getting
kind of large.

Bug: 81915
Change-Id: Ie60c1ba3487be07964470249c3dc80c54cf7d954
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612021
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>
API-Review: Alice Neels <neelsa@google.com>