[fdf] Add DFv1 / DFv2 interop integration test
This integration test verifies that we can load a DFv1 driver through
the compatibility driver in the middle of a chain of DFv2 drivers. This
shows that a DFv1 driver can be both a child and a parent of a DFv2
driver.

Fixed: 86512
Bug: 33183
Change-Id: If37ca2f2075b4a4e052c5c6de78a53e4ae8df8d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613548
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>