[starnix] Add vulkan initialization code to starnix
This code creates vulkan instances and devices, and sets constraints on
a shared buffer collection.

The code will be tested by running the unit tests for minigbm, but those
tests are not currently available in a form that can be run on starnix.
The change has been tested locally, and eventually the test binary will
be run as part of the starnix test suite.

Bug: 89887

Change-Id: I9f40d70d16306c822f6228520e7579bcfc9af518
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613302
Reviewed-by: David Reveman <reveman@google.com>
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>