[kernel][vm] Package pager writeback core tests
The pager writeback core tests only run in the unified mode on bringup
builds. Package the test so that it can also run as a standalone binary
in core builds.

Ideally we would want to have the next vDSO available for standalone
tests on the bringup build, which is tracked in fxbug.dev/89597.

Bug: 89597
Bug: 63989
Change-Id: I012ca2bb88596fe1971d3534cb4c1979ffc3d22f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/608543
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>