[lowpan] Move types out of lowpan-spinel-driver to lowpan-driver-common
In the process of developing a new LoWPAN driver that directly uses
LoWPAN, it seems I need a lot of the types that were previously only
present in `lowpan-spinel-driver`. This CL moves these types to the
shared `lowpan-driver-common` crate so that they may be used by the
future `lowpan-ot-driver`.

Change-Id: I5fff40fcad19263bfb2cf01b1f5e0f8f6f4fe13c

Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605984
Commit-Queue: Robert Quattlebaum <rquattle@google.com>
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>