[driver_host] Fix flake
Removing the vnode seems to result in removing a reference from the vfs
to the device, resulting in a use after free when the vfs is unable to
finish cleaning up access to the device prior to the device being freed.
The UnbindDevice call earlier in the test will result in a call to
DriverHostContext::DriverManagerRemove which will reset the device vnode
at the appropriate time.

Bug: 86267
Multiply: driver-host-tests
Change-Id: Idf688cb9cdbc4fdd375f69b6d74353ac4fb0bcda
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614742
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>