[fx][add-update-source] Set the update channel
When using `fx add-update-source`, we register the repository and change
the fuchsia.com alias to point at the repository, but we do not change
the update channel to point at this repository. This can lead to
confusing error during the transition to the new repository server when
installing an OTA. If a device has both the ffx and pm repositories
registered, but only one of the servers is online, we can get this
error:

```
% fx shell update check-now --monitor
Checking for an update.
State: CheckingForUpdates
Error: Update failed: ErrorCheckingForUpdate
```

This can happen if the device is trying to install an OTA from the old
offline repository.

To fix this, this patch changes `fx add-update-source` to always set the
update channel after registering a repository. This makes sure that the
device will take an OTA from the right place.

Change-Id: I479c1e8fb0ae433657c5f6b518694eb2388d7041
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613903
Reviewed-by: Dan Johnson <computerdruid@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>