[rust][time] Migrate ClockUpdate to zx_clock_update_args_v2_t.
With v2 there are more fields and we don't want to support all
combinations. The typestate builder pattern works really nicely for this
since it ensures code that sets invalid combinations (or even sets the
same property twice) will fail to compile.

This means going to a `ClockUpdateBuilder` that is generic over the
state of each property and which may be converted to an immutable
`ClockUpdate`. This requires a small change in all callers.

Note that the advantage of zx_clock_update_args_v2_t is the ability to
specify an absolute time, but this CL doesn't take advantage of that
ability yet, all clients are converted to use "approximate_value" which
is equivalent to the v1 functionality. We temporarility allow the
combination of approximate time and rate during the transition but
intend to disallow this in the future.

Bug: 65726
Change-Id: Ia2387a3d10a2b31a45f7269c43bec57393ba6df4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613661
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>