[archivist] More informative error type
Refactors all usage of `anyhow::Error` that isn't for tests into its own
error type. Additionally, the inspect collector used to fail if a single
file in the directory failed to be read. This shouldn't be the case and
this CL aligns the behavior of reading from a file with the behavior of
reading from a protocol.

Fixed: 51237

Change-Id: Icbedf3cf90f54735470c11b8ca8d1b4b21780693
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614941
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>