[hub] Remove 'deleting' directory
This CL re-lands the changes introduced in fxr/605042 by only including
the minimal change set needed to remove the `deleting` directory.

The `deleting` directory is removed from struct `Instance` and no longer
serves as a holding directory for destroyed instances between the
`on_destroyed_async` and `on_purged_async` event handler functions.
References of the `deleting` directory are removed from tests.

The reason for this change is because there are no current use cases of
the `deleting` directory. Furthermore, this directory uses the instance
ID of monikers which is in the process of being deprecated (x-ref
fxr/605042).

Bug: 87405
Change-Id: I58da854b73981838799004f1c89002e435619c0f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614983
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>